% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ini.R
\name{read.ini}
\alias{read.ini}
\title{Read and parse .INI file}
\usage{
read.ini(filepath, encoding = getOption("encoding"))
}
\arguments{
\item{filepath}{.INI file to parse}

\item{encoding}{Encoding of filepath parameter, will default to system
encoding if not specifield}
}
\value{
List with length equivalent to number of [sections], each [section]
will be a named vector
}
\description{
Read and parse .INI file
}
\seealso{
\code{\link{write.ini}}
}

