% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ini.R
\name{write.ini}
\alias{write.ini}
\title{Parse list structure to .INI file}
\usage{
write.ini(x, filepath, encoding = getOption("encoding"))
}
\arguments{
\item{x}{List with structure to be write at .INI file.}

\item{filepath}{.INI file to write}

\item{encoding}{Encoding of filepath parameter, will default to system
encoding if not specifield}
}
\description{
Parse list structure to .INI file
}
\examples{
\dontrun{
## Create a new list holding our INI
newini = list()
newini[[ "Section Name" ]][[ "Key name" ]] = "Key value"
newini[[ "Section Name 2" ]][[ "Key name" ]] = "Key value"
## Write structure to file
write.ini(newini, "written.ini")
}


}
\seealso{
\code{\link{read.ini}}
}

